requirements.txt
langgraph>=0.2.70
langchain-core>=0.3.0
langchain-mistralai>=0.2.0
python-dotenv>=1.0.1
pydantic>=2.7.0
openinference-instrumentation-langchain>=0.1.28
opentelemetry-sdk>=1.30.0
opentelemetry-exporter-otlp>=1.30.0
arize-phoenix>=11.0.0
